+++ stories/Eltanin/model/eltanin/locality/scrap.q1.cpp
@@ -104,7 +104,7 @@ namespace eltanin::locality {
                 prevOri = glm::normalize(step * bodyPose.rotation);
             }
             with<phys::rigid::Solid>::extend(context, body, phys::rigid::Solid::Quantum{
-                .center = phys::Particle{phys::Matter{.position = dvec3{bodyPose.position}, .mass = mass, .temperature = temperature, .cohesion = cohesion}, dvec3{bodyPose.position} - dvec3{linear * float(phys::Settings::fixedStep)}, vec3{0.0f, 0.0f, 0.0f}},
+                .center = phys::Particle{phys::Matter{.position = dvec3{bodyPose.position}, .mass = mass, .temperature = temperature, .cohesion = cohesion}, dvec3{bodyPose.position} - dvec3{linear * float(phys::Settings::fixedStep)}, dvec3{0.0, 0.0, 0.0}},
                 .prevOri = prevOri,
                 .forceAngular = vec3{0.0f, 0.0f, 0.0f},
                 .kind = phys::rigid::Solid::Kind::box,
